mt7612u: fix the defects under the measurements
## Problem

Review of #412 found bugs that the hardware measurements sat on top of. Two
of them change what the driver puts on air or hands to a caller, one is a
use-after-free during teardown, and one made the "standalone library" claim
false.

## Change

Ten confirmed findings, each with the fix and its evidence.

**The public header declared nine functions that had no definition anywhere**
(`mt7612u_open/_close/_keep_detached/_set_channel/_set_chainmask/_start/
_stop/_asic_version/_mac_addr`). The bringup tool builds because it calls
internals directly, so nothing caught it. They are now defined, and
`tests/api_link.c` takes the address of all twenty public entry points while
including only the public header, so the link fails if a declaration ever
loses its definition again. `set_channel` also disagreed with the internal
signature on `unsigned` vs `uint8_t`; it is `uint8_t` now.

**The RX L2-pad fold moved a fixed 24 bytes.** L2PAD is only ever set when
the header is not 4-aligned, i.e. 26 bytes (QoS) or 30 (4-address), so the
last two header bytes were left behind and overwritten by the pad. On a QoS
frame those two bytes are the QoS Control field, which means every TID and
ack-policy reached the callback as zero. It now moves the real header
length, as mt76x02_remove_hdr_pad() does.

**Header length treated only RTS and PS-Poll as 16-byte control frames.**
BlockAckReq, BlockAck and both CF-End subtypes are also 16, so on TX the L2
pad was inserted ten bytes in, inside the frame. Replaced with a port of
ieee80211_hdrlen(): control frames are 16 by default and only CTS and ACK
are 10. That is the same function mt76 reaches via
ieee80211_get_hdrlen_from_skb() on both sides, so TX and RX now share one
implementation instead of two that could drift.

Both are covered by `tests/frame_shape.c`, which also carries a negative
control: it redoes the old fixed-24 fold and asserts the QoS Control really
is destroyed, so a passing test says something. Reverting either fix makes
it fail - the RX one reports `QoS Control zeroed by the pad fold: aa aa`.

**mt_async_stop() cancelled only RX, then freed every transfer.** On a
wedged chip - the case this driver's own notes describe - TX URBs never
complete, and libusb owns a submitted transfer until its callback runs. Now
both rings are cancelled, both are waited on with the event thread still
running, and if either still has transfers outstanding at the deadline the
ring is deliberately leaked rather than freeing memory the kernel may still
write into. Each slot names its own ring instead of dev->a, so a leaked
ring's late completions cannot touch a replacement one. The `fail:` path no
longer joins a zeroed pthread_t. `running`, `rx_active` and both inflight
counters move under the mutex - volatile orders nothing and makes no
read-modify-write atomic - and the statistics move with them behind
mt_async_stats(), because the harness was reading them live off the struct.
A caller blocked waiting for a TX slot now gives up when the ring stops.

**mt_rr() returned ~0u on failure and mt_rmw() consumed it unchecked**, so
one exhausted EP0 read wrote 0xFFFFFFFF | val into MT_WLAN_FUN_CTRL,
MT_MAC_SYS_CTRL or the BBP AGC block. 0xffffffff is a real value on this
part (MT_MAC_CSR0 reads it while the core comes up), so it cannot double as
a sentinel: mt_rr_chk() reports failure separately, mt_rmw() refuses to
write after a failed read, and mt_wait_for_mac() distinguishes "not ready"
from "no transport". mt_poll() polls against a real deadline and aborts on
a read failure - one register access can cost VEND_RETRIES * timeout, so
counting sleeps let a caller asking for 200 ms block for seconds. The
control timeout drops to mt76's own 300 ms.

**mt_wr_copy() rounded len up and then memcpy'd from the caller's buffer**,
over-reading whenever len % 4. Latent at the current call sites (8 and 32
bytes); it now rounds the transfer up and zero-fills instead.

**Two bringup gates enabled MAC RX and never drained EP 4** - the exact
wedge pattern documented three paragraphs above them, and one of them then
sat through two 200 ms sleeps and a channel switch. Both start the ring
before the receiver. mt7612u_start() enables RX only when an RX ring is
already running, and the header says why.

**Radiotap DBM_TX_POWER was parsed and silently dropped.** It now says so
once per process, pointing at the two knobs that do work.

**The mt76 TSF bug claim now carries its citation** - mt76x02_usb_core.c:155-158,
where tsf = (u64)dw0 << 32 | dw1 feeds only the dev_dbg() on the next line,
which is why the order has survived upstream.

Also fixed while in send_packets(): the radiotap parse moved into the
selection pass. A frame the build pass could still reject would break the
chain it was building, since NEXT_VLD and the single trailing zero word are
assigned by position - dropping whichever frame happened to be last left the
transfer unterminated.

## Verification

`make -C src/mt7612u check` - twenty public entry points resolved,
frame_shape PASS. Both new tests were mutation-tested: removing one
definition, and reverting each frame-shape fix, each makes them fail.
No hardware and no privileges needed. Clean build, no warnings.

`CMakeLists.txt` is still untouched.

+++ src/mt7612u/include/mt7612u/mt7612u.h
@@ -85,10 +85,12 @@ void mt7612u_close(struct mt7612u_dev *dev);
 void mt7612u_keep_detached(struct mt7612u_dev *dev, int keep);
 
 /*
- * Channel + width. Issues CMD_SWITCH_CHANNEL_OP and the firmware calibration
- * burst, so it is not cheap - it is a setup call, not a per-frame one.
+ * Channel + width. `chan` is an 802.11 channel number. Issues
+ * CMD_SWITCH_CHANNEL_OP and the firmware calibration burst, so it is not
+ * cheap - it is a setup call, not a per-frame one. Only 20 and 40 MHz are
+ * implemented; 80 MHz is silicon-capable but the width maths is not ported.
  */
-int mt7612u_set_channel(struct mt7612u_dev *dev, unsigned chan, enum mt7612u_bw bw);
+int mt7612u_set_channel(struct mt7612u_dev *dev, uint8_t chan, enum mt7612u_bw bw);
 
 /* Absolute TX power base, dBm. Per-frame trim is mt7612u_tx_rate.power_adj. */
 int mt7612u_set_txpower(struct mt7612u_dev *dev, int dbm);
@@ -96,7 +98,13 @@ int mt7612u_set_txpower(struct mt7612u_dev *dev, int dbm);
 /* 0x202 = 2T2R (default), 0x101 = 1T1R. Global; takes effect at next channel set. */
 int mt7612u_set_chainmask(struct mt7612u_dev *dev, uint16_t chainmask);
 
-int mt7612u_start(struct mt7612u_dev *dev);  /* enable MAC TX+RX */
+/*
+ * Enable the MAC. TX always; RX only if mt7612u_rx_start() is already
+ * running. That condition is not a convenience: with the receiver on and
+ * nothing draining the bulk-IN endpoint, this part wedges below the USB
+ * level and no software reset recovers it - only a physical replug.
+ */
+int mt7612u_start(struct mt7612u_dev *dev);
 int mt7612u_stop(struct mt7612u_dev *dev);
 
 /*
